v1.0.0 Lines now start immediately after the pipe | instead with a single space after the pipe.
This was done because over the years I've found that many editors (e.g. vscode) have a "trim trailing whitespace" feature that is usually not context aware and instead simply removes all trailing whitespace on every line.

This is probematic when used with this library because previously you would not have been able to have blank lines work at all because your editor would strip out the single character needed to signal it is supposed to be a blank line.

Before:

```js
  | foo
  | <---- there is a single empty space here
  | bar
```

After

```js
  |foo
  |<----- no empty space, but still a blank line
  |bar
```